requirements.txt
aniso8601==4.0.1
Click==7.0
flasgger==0.9.1
Flask==1.0.2
Flask-RESTful==0.3.6
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mistune==0.8.4
pytz==2018.7
PyYAML==3.13
six==1.11.0
Werkzeug==0.14.1
